+++ ElvUI_MerathilisUI/Core/Changelog/6.64.lua
@@ -1,8 +1,18 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[664] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	RELEASE_DATE = "20.05.2025",
+	FIXES = {
+		"[Notification]: Fixed full bag notification",
+	},
+	NEW = {
+		"[Focuser]: Added new option to shift+click a unit to set them as your focus",
+	},
+	IMPROVEMENTS = {
+		"[Misc]: Added back old ElvUI Role Icons",
+		"[RBReminder]: Added more food, rune ids. Thanks @vociferoy from Discord",
+		"[AutoButtons]: Added more food and flask ids",
+		"[GameMenu]: Show Mythic+ infos on the GameMenu by default again",
+		"[Installer]: Some minor changes to the installer",
+	},
 }