+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ### Changes ###
 
-  * [World Map]: Fix loading of EventTracker >.>
+  * [Notification]: Fixed full bag notification
+  * [Focuser]: Added new option to shift+click a unit to set them as your focus
+  * [Misc]: Added back old ElvUI Role Icons
+  * [RBReminder]: Added more food, rune ids. Thanks @vociferoy from Discord
+  * [AutoButtons]: Added more food and flask ids
+  * [GameMenu]: Show Mythic+ infos on the GameMenu by default again
+  * [Installer]: Some minor changes to the installer